Improve pre-commit hook for formatting
Instead of using the pre-commit-npm-package we now use a combination of husky and pretty-quick.

The pre-commit-package is not capable of repository structures, where the .git directory and the package.json file are not on the same level.
That is the case in our repo (as the package.json for the bp-gallery React-App is located under projects/bp-gallery).

Husky is Commit-Hook-package which is capable of that.
See https://typicode.github.io/husky/#/?id=custom-directory
for more details on that.

Using pretty-quick (as the name says) helps us saving a lot of time.
Instead of invoking a script which formats all our source files (via yarn format which we configured prior for example),
it will execute the autoformatting via prettier only on STAGED files.
That means formatting will only take place on src-files that are about to be committed.