Refactor testUtils' renderRoute
As we now got a Redirect in our route definitions we can't just use a 
MemoryRouter in routing tests anymore. Instead we use the more flexible 
BrowserRouter.

Also in that way we avoid the setTimeout in the routing-tests. Timeouts 
are evil in tests because they will always make them pass, even if you 
assert false for example :P