Feat/frontend
* started frontend

* Added (empty) Components for PictureView

* Create basic file structure and routes

* Disabled Linter function: "no-floating-promises"
Reason: Calling async functions in useEffects() leads to linting errors while there is no alternative.
https://github.com/facebook/react/issues/14326

* First draft to display pictures

* Sample code for browse view

* Fix merge issues and tests

* Fix linting

* graphql queries and hierarchical browsing

* linter

* implemented basic features of details and comments view, started styling of pictureView

* fixed linting errors

* New scroll bar

* Added more styling to picture view

* Further styling of PictureView + Added thumbnailMode of picture view. Changed linting rule: no-unsafe-call

* fixed mistake in test

* Show image grid and add search bar

* Improve and fix pictureView-test

The chain of await/asyncs and acts was a bit overkill right here and 
also didn't actually work.
Checkout [1] for a better understanding of async testing.
Also renamed the test file from .ts to .tsx to support jsx in the future 
here.

[1] https://testing-library.com/docs/dom-testing-library/api-async

* Refactor testUtils' renderRoute

As we now got a Redirect in our route definitions we can't just use a 
MemoryRouter in routing tests anymore. Instead we use the more flexible 
BrowserRouter.

Also in that way we avoid the setTimeout in the routing-tests. Timeouts 
are evil in tests because they will always make them pass, even if you 
assert false for example :P

* Get rid of a quite meaningless test in app.test.tsx

The 'Navigation bar is always visible' test is quite 
meaningless/unlogical for following reasons:
- it is simply not true as the navbar isn't visible in all routes (e.g. 
the demo prototypes)
- in the pictureView and galleryView tests some things of the navbar get 
queried, so it is unnecessary here
- and most importantly (together with the first point) the use of 
'setTimeout' has hidden the fact that the test is actually failing.

* Overall improvement of the tests

- Use the `expect(container).toHaveLength(...)` matcher instead of 
`expect(container.length).toBe(...)`
- Query on the rendered component instead of the whole document

* refactoring of pictureView and small css additions

* reverted refactoring-change in PictureView

* removed unused import

* Refactor browse view

* Update projects/bp-gallery/src/ApiConnector.ts

Find ick gut

* Update projects/bp-gallery/src/components/NavigationBar.tsx

* Update projects/bp-gallery/src/components/TopBar.tsx

* Include pr review suggestions

* Remove console log

* Reverse supposed fix

Co-authored-by: Hendrik Woelert <hendrik.woelert@online.de>
Co-authored-by: Benedikt Vidic <benedikt.vidic@student.hpi.uni-potsdam.de>
Co-authored-by: Olliwehr <oliver.hess@student.hpi.uni-potsdam.de>
Co-authored-by: TheGrayStone <lissner.nils@liwe-design.com>
Co-authored-by: philbow61 <philip.raetsch@gmail.com>