Add tests for the PictureView (together with its API mocks)
It is located in a 'test'-directory next to the location of the Component itself. Next to the test there is a file called 'mocks.ts' in which the mocks for the GraphQL-API (and other test data) is put.

The PictureView tests also make use of the describe/it pattern and just test the 'compositon unit' (so other components in the sub-tree are also mocked).

With the help of the GraphQL-mocks we can test the following:
- what happens when api-data is ready (needs waitFor after the render)
- what happens when data is loading (no waitFor needed)
- what happens when a network-error occurs (needs waitFor)
- what happens when a GraphQL-error occurs (needs waitFor)