Adjust APIConnector-generation to the v4 way
The big problem is that Strapi v4 doesn't export a separate schema file anymore. So we now need to query the '/graphql' endpoint with an introspection query instead in order to retrieve the schema.

The drawback with this approach it that the server needs to be available.
In order to also make it possible to run the `yarn generate-api` command when the server is unavailable, we now also manage a local schema file as a fallback. That is also important for executions in the CI-environment.

Note that we need a custom remote loader function in order to not exit the whole codegen-process when the server is unavailable, but instead use the local copy as a fallback. The file gets ignored by the linter for now.

Ressources:
- https://www.graphql-code-generator.com/docs/config-reference/schema-field#custom-schema-loader
- https://gist.github.com/craigbeck/b90915d49fda19d5b2b17ead14dcd6da#gistcomment-3204270