Use different "contains" filter for the v4
In Strapi v3 we used the "contains" filter for loading stuff based on the typed search-term.
This filter was case-*insensitive* (see https://docs-v3.strapi.io/developer-docs/latest/development/plugins/graphql.html#query-api). For a case-sensitive filter "containss" (with a "s" suffix for "sensitive") had to be used.

In Strapi v4 tho, the default "contains" is now case-*sensitive* (see https://docs.strapi.io/developer-docs/latest/developer-resources/database-apis-reference/graphql-api.html#filters).
So in order to not break the case-insensitve matching of the typed search-term, now "containsi" (with an "i" suffix for "insensitive") needs to be used instead.

uff...