Fix bugs that were not detected when merging dev/search
When searching for keywords/descriptions with spaces and other stuff included, i noticed that we get no pictures back.
Thats because the search params werent decoded from the URI schema before putting them into the API filters object.
But it is quite possible, that this only populates to the Strapi v4 now.

Also removed the 'value' prop of the searchbar as this is longer needed.