Implement custom React-Hook `useFlatQueryResponseData`
As stated in the prior commit, I thought about extracting the `flattenQueryResponseData` function into its own React-Hook.
This way I can now use `useMemo` directly in my exported function, so the stateful components that will need flattened response data don't need to wrap a `useMemo` call around the flatten-call themselves. :)

In the first place, memoization was needed to fix an infinite render-loop. The function always returns a new object (different location in memory) even though the input data was the same object as for the previous call. That resulted in those render-loops when the returned data was used as an useEffect-dependency.