Refactor the additional Flat types
Before only a part of the content types had a flattened equivalent in this file, now every content type that is present at the current point has one.

Basically, first an equivalent is created that omits the relational fields of a type. These get added back in with an equivalent of the relational partner that is also flattened. So in that way we achieve the flattening of types on the top-level (id together with attributes) and also on the next level of relational fields.

Also adds custom flat types for the two queries that make sub-queries to the pictures content types in order to retrieve a thumbnail.