Refactor queries and shape of CommunityView
To only get the related tags (of a Category) that contain "new" pictures, we now execute a light-weight query to get the IDs of said tags and then we execute the prior existing to get the actual info of the current tag.
After that the potential related tags get filtered by the IDs.

Also fixed that the PictureScrollGrid now actually respects the publishingDate of pictures when loading them.

Co-authored-by: rfrx <rieke.freund@student.hpi.de>