Fix comment posting (#48)
* Fix comment posting

Without the retrieved picture id, comment posting isn't possible anymore!

* Add 'id' field in PictureInfo-Mocks

* Write tests for the NewCommentForm

With these we should notice, if comment posting isn't working anymore.
Btw it is the first time we mocked a mutation in tests.

Behaviorwise, the NewCommentForm now only resets the input fields and shows the success alert if the mutation really completed.

* Use AlertWrapper for the Alerts we show after commenting

Unfortunately i couldn't make the tests work, as the `onCompleted` and `onError` callbacks of the mutation just don't get called in the tests :(
But i tested it locally with a dev version of the app, there they get called.
Im confusion.

Also used the AlertWrapper in tests.

* Re-introduce the tests for the comment posting

The big problem was the faulty time mocking before, glad Hendrik helped me find the bug! <3

Note that the extra helpers file is in order to make the DateTime mockable.