Add zoom capabilities to picture view (#50)
* Add zoom capabilities to picture view

Honestly, a lot happened here. Mainly, you can now use the zoomwrapper to zoom into the picture. Furthermore, to circumvent issues with the scroll behavior, the "intelligent" scrolling from before was replaced so that now you can open the side/bottom-bar with a button. This makes some parts a lot cleaner and may be better for the end user as well.

* Add translations

* Refactor DEFAULT_ZOOM

* Apply review comment

* Fix picture view navigation bar tests

* Fix PictureInfo tests

* Add zoom wrapper tests