Only fire a single PictureTitleUpdateMutation
In order to only send one request to the server. To make that work, you can now specify in the meta key of the title update which other pictures (based on the one that is currently shown) should receive the same title update.
That is possible via two ways:
- set `updateAllRelatedPictures` to true, so all pictures related to the same (old) title as the current pictures get updated
- and/or also specify picture ids directly in the `updateAdditionalPictures` array

Those two options are wrapped in a `meta` key which also replaces the role of the `customUpdate` key.

In the lifecycle we trigger a recursive update (which each triggers the whole lifecylce itself) for the additional pictures via the sequential processing of promises. See https://css-tricks.com/why-using-reduce-to-sequentially-resolve-promises-works/ for more details on that.