Implement the recursive merging of verified/unverified data
Note: still in draft state

In analogy to the flattenQueryResponseData function there now is a recursive function to merge the data of verified and unverified relations.
When it comes to a n:m relation it will output an array first filled with the verified entities and the unverified ones afterwords.
When it comes to a 1:n relation it will always use the verified entity if there is one.

Also the function will add a verified flag to each entity it processed.