Draft: Adapt spiked picture beforeUpdate lifecycle for field editing
This is a first version of the beforeUpdate lifecycle.
For following fields of a picture there now is a first-level (without additionalPictures) editing possibility:
- descriptions
- time_range_tag
- person_tags
- location_tags
- keyword_tags

The lifecycle is heavily dependent that in the stringified json, there is a verified field for the types that got some verified handling and also on the the prevId field, that signals (when set) that the prior tag should be updated instead of creating a new one with the given data.

As well as the support for the bulk-editing (updateAdditionalPictures) there is also a TODO on the line that early-returns when it detects the Admin-Panel triggered the update (then we shouldn't trigger any special handling).