Implement the merge of a description with another existing one
This is a special case when the frontend defines that the previous description should be updated completly with a new text, but there is already an existing description with that text.

In that case, the handling now will relate all pictures with the previous description to the already existing one and un-relate all from the previous one. This will then result in a cleanup of the previous description as its now unrelated.