Scan pictures directly into the software
Explanation: In order to use this feature, a WebSocket Service has to run on the curator's system (python script). To make usage on Windows easier, I added a .zip file which contains a packed version. You can execute the exe file inside to start the service.

When the service is running, communication via websockets should allow any WIA-compliant scanner  to send pictures to the system.