Setting the errorPolicy to 'all' where needed
As errors are now globally caught, in order to prevent calls of `onCompleted` hooks even though an error occurred, we need to manually set the errorPolicy as strict as possible.

Unfortunately, this is not globally doable, as the defaultOptions are not read from the ApolloClient but instead are generated via the GraphQL Codegen.