Globally set the ErrorPolicy to 'all'
In order to that we configure the codegen plugin to set this as a default option.

Afterwards we then also need to manually replace sth. in the generated file (with the package replace-in-file), in order to tell the TS Compiler that the string 'all' is indeed assignable to the String Union Type 'all' | 'ignore' | 'none'.
A bit hacky but it works (hopefully :D)