Include archive tags in the search (#124)
* Include Archive-Tags in the All-Search

It works like a regular search term of the All-Search. Just search for sth. like "<your normal search term> <your archive search term>".

* Make archive tags searchable in the Advanced Search

* Make archive tags clickable also for non-curators

In that case a new tab gets opened with the search view for that archive tag.
Thats similar to how we handle that for the other tags.