Remove LanguageDetector module
Since we currently only support German anyway, this is useless.
Furthermore, it actually produces some unwanted results:
When the LanguageDetector detects en-US (e. g. on my system),
i18next falls back to de, since that's the only language we support.
But for formatting, it still passes en-US to the Intl.*Format.
This results in numbers being formatted in the US style,
while the actual translations are German
(e. g. 1,500 Bilder instead of 1.500 Bilder).