276 improve i18n usage (#278)
* Use plurals and formatting for pictureCount and moreThanPictureCount

* Use plurals and formatting for links copying and pasting

* Use formatting for really-decline-comment-descendants

* Remove LanguageDetector module

Since we currently only support German anyway, this is useless.
Furthermore, it actually produces some unwanted results:
When the LanguageDetector detects en-US (e. g. on my system),
i18next falls back to de, since that's the only language we support.
But for formatting, it still passes en-US to the Intl.*Format.
This results in numbers being formatted in the US style,
while the actual translations are German
(e. g. 1,500 Bilder instead of 1.500 Bilder).

* Fix https://github.com/hpi-swa-lab/BP2021RH1/issues/276#issuecomment-1440426332

* Fix typo

* Fix merged tests

* Fix the fix

* Fix test

---------

Co-authored-by: Oliver Schulz <64468276+olschulz@users.noreply.github.com>