Remove unnecessary factor 2
It was introduced in c6c446975bc76aba354d5517a1fd91fdb17f73be to prevent the
image from moving twice as fast as the pointer. The original
problem was that in StrictMode, the updater function passed to a
setState, in this case setZoomLevel, is called twice to ensure
side-effect freedom. This resulted in setViewport being called twice,
moving the viewport twice the desired amount. In
9871e2f521d7518e773f2abed47cea57e057ba8d, the nested setState call was
removed, so now the leftover factor 2 has to be removed as it produces
unwanted half-as-fast panning when StrictMode is off.