594 further improve loading time for hierarchy (#596)
* Regenerate strapi typing

* Fix PopulateType

* Add custom resolver for getAllLocationTags

This is a performance optimization. Strapi's graphql resolvers run a
separate query for each location tag to fetch their child and parent
tags, i. e. for 500 location tags, 1000 database queries would be
executed. By using the entityService directly, we avoid the many
database queries and replace them with a single one.

This also removes the child tags fetch and only fetches the parent tags
as the client builds a graph from them anyway.

* unnecessary condition

* Stringify ids in getAllLocationTags

---------

Co-authored-by: LinoH5 <lino.hellige@student.hpi.uni-potsdam.de>