Permissions enhancements (#605)
* Add sticky headers to PermissionsView

* Replace plural header archives with singular for archive specific permissions

* Improve filter input visibility

* Fix save state not showing when user can't edit picture

* Disable map when user can't update location coordinates

* Refactor LocationManagementDialog buttons

* Only show editable archives in ArchivesView

* Fix Unauthorized error on MyAccountView when user can't run getUser

* Remove ability to login via resetPassword

Also remove jwt token from changePasswort response.

* Fix error handling again

* Extract location specific permissions into own section

* Improve translation for withoutArchive

* Clarify getUser and updateUser translations

* Split off updateCollectionParents from updateCollection

* Adapt ListItemButtons for users according to permissions

* Translate error messages in QueryErrorDisplay

* Default to pictures and texts when logged in

* Extract PresentationModeButton

* Add noDistractionModeStyle to PresentationModeButton

* Batch addPermission and deleteParameterizedPermission

This is to mitigate a bug/behavior with refetchQueries, where
getParameterizedPermissions is refetched when the first mutation comes
back, but the refetch triggered by later mutations somehow does not run.
Batching the mutations solves this problem, since the refetch can only
start after all of the mutations have come back, which they can only do
together, since they are a batch (currently, there are never more than
50 mutations run at the same time). It also reduces network load as a
side benefit.

* Manually refetch permissions

This is the second part of the fix for the problem described in the
previous commit. Batching alone does not help when BooleanParameters
change, since they rely on the correct order of delete and add, so the
refetch triggered by the delete overshadows the refetch triggered by the
add. Refetching manually solves this problem, since a refetch is only
triggered after a logical action is completed.

Batching is kept for the network load aspect.

* Extend public preset to basic preset

* Make section titles toggle accordions instead of checkboxes

* Extract SaveStatus

* Add SaveStatus to PermissionsView

* Fix tests

* Fix a5beb0a298298ea0fd4b4a3721bd0ee473b4fd4c for delayed auth loading

* Fix text filter tests

* Serialize picture uploading

* Rename working to isWorking

* Readd missing copy of users array