Fix PictureSequenceInfoField sometimes showing no pictures
Before, when switching between two pictures, where the first is part of
a  sequence, the the other pictures in the sequence are not shown when
switching back to the first. The exact reason is unclear, however print
debugging has shown that during switching, the correct array of pictures
is displayed first, only to be replaced by an empty array shortly after.
Since this commit fixes the issue by changing the fetchPolicy from
'cache-first' (default) to 'cache-and-network', the problem is likely
related to Apollo's caching behavior.